requirements.txt
asgiref==3.11.1
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.2
django-auditlog==3.4.1
django-configurations==2.5.1
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-phonenumber-field==8.4.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
oauthlib==3.3.1
packaging==26.0
phonenumbers==9.0.23
pillow==12.1.1
pluggy==1.6.0
psycopg2==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.16.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
urllib3==2.6.3
